tests/performance/requirements.txt
grpcio==1.24.3
numpy==1.18.2
tensorflow-serving-api==1.15.0
cpplint==1.4.5
